fix maximum size

+++ Backend/src/controller/SetupProjectApi.js
@@ -1,17 +1,43 @@
 const util = require('util');
 const download = require('download-git-repo');
 const exec = util.promisify(require('child_process').exec);
-const { getProjectSize } = require('./GetInfoApi')
+const { getProjectSize } = require('./GetInfoApi');
+const axios = require('axios');
 
-exports.CloneProject = (req, res) => {
+exports.CloneProject = async (req, res) => {
     request = req.body;
-    var repo = request.url;
-    var username = request.username;
-    var name = request.name;
+    let repo = request.url;
+    let username = request.username;
+    let name = request.name;
 
     console.log("CLone GIT: ", repo, username, name);
-    exec(`mkdir -p ../data/result/${username}/${name}`);
 
+    let user_repo = repo;
+    if (repo.includes("https://")) user_repo = repo.substring(19, repo.length);
+    else user_repo = repo.substring(11, repo.length);
+
+    console.log(user_repo);
+
+    let size = 0;
+    let private = false;
+    await axios.get(`https://api.github.com/repos/${user_repo}`)
+        .then((res) => {
+            console.log(res);
+            size = res.data.size;
+            private = res.data.private;
+        });
+
+    console.log(size);
+    if (size > 100000) {
+        res.status(404).send({ size: size * 1000 })
+        return;
+    }
+    if (private) {
+        res.status(404).send({ message: 'Private' })
+        return;
+    }
+
+    exec(`mkdir -p ../data/result/${username}/${name}`);
     download(`direct:${repo}`, `../data/${username}/${name}`, { clone: true }, async (e) => {
         if (e) {
             exec(`rm -rf ../data/result/${username}/${name}`);
@@ -19,9 +45,8 @@ exports.CloneProject = (req, res) => {
             console.log('Error: ', e);
         }
         else {
-            // await res.status(200).send({ message: 'Success' });
-
-            await getProjectSize(req, res);
+            await res.status(200).send({ size: size * 1000 });
+            // await getProjectSize(req, res);
             console.log('Success');
         }
     });
@@ -30,8 +55,8 @@ exports.CloneProject = (req, res) => {
 exports.UploadProject = async (req, res) => {
     request = req.body;
     file = req.file;
-    var username = request.username;
-    var name = request.name;
+    let username = request.username;
+    let name = request.name;
 
     console.log("Upload Project: ", file.originalname, username, name);
